Move adapter step entrypoints to adapters/steps/
Move trigger, loader, and reloader step files from
extractors/{ebsco,oai_pmh}/steps/ to the top-level
adapters/steps/{ebsco,oai_pmh}/. Also moves
transformers/transformer.py to adapters/steps/transformer.py.

This consolidates all step entrypoints under a single steps/
directory, namespaced by adapter type, matching the pattern
used by ingestor/ and graph/.
