--- catalogue_graph/src/adapters/extractors/ebsco/steps/__init__.py
@@ -1,1 +1,0 @@
-# Steps package
