--- catalogue_graph/src/adapters/extractors/oai_pmh/steps/__init__.py
@@ -1,1 +1,0 @@
-"""Generic OAI-PMH adapter step implementations."""
