+++ catalogue_graph/tests/adapters/extractors/oai_pmh/test_reloader.py
@@ -15,8 +15,8 @@ from _pytest.monkeypatch import MonkeyPatch
 from pyiceberg.table import Table as IcebergTable
 
 from adapters.extractors.oai_pmh.runtime import OAIPMHRuntimeConfig
-from adapters.extractors.oai_pmh.steps.loader import LoaderRuntime
-from adapters.extractors.oai_pmh.steps.reloader import ReloaderRuntime, handler
+from adapters.steps.oai_pmh.loader import LoaderRuntime
+from adapters.steps.oai_pmh.reloader import ReloaderRuntime, handler
 from adapters.utils.adapter_store import AdapterStore
 from adapters.utils.window_store import WindowStore
 from adapters.utils.window_summary import WindowSummary
@@ -111,7 +111,7 @@ class TestHandler:
             ]
 
         monkeypatch.setattr(
-            "adapters.extractors.oai_pmh.steps.reloader.build_harvester",
+            "adapters.steps.oai_pmh.reloader.build_harvester",
             lambda event, runtime: SimpleNamespace(harvest_range=mock_harvest_range),
         )
 
@@ -179,7 +179,7 @@ class TestHandler:
             ]
 
         monkeypatch.setattr(
-            "adapters.extractors.oai_pmh.steps.reloader.build_harvester",
+            "adapters.steps.oai_pmh.reloader.build_harvester",
             lambda event, runtime: SimpleNamespace(harvest_range=mock_harvest_range),
         )
 
@@ -252,7 +252,7 @@ class TestHandler:
             raise RuntimeError("OAI-PMH endpoint unavailable")
 
         monkeypatch.setattr(
-            "adapters.extractors.oai_pmh.steps.reloader.build_harvester",
+            "adapters.steps.oai_pmh.reloader.build_harvester",
             lambda event, runtime: SimpleNamespace(
                 harvest_range=mock_harvest_range_error
             ),
@@ -310,7 +310,7 @@ class TestHandler:
             )
 
         monkeypatch.setattr(
-            "adapters.extractors.oai_pmh.steps.reloader.build_harvester",
+            "adapters.steps.oai_pmh.reloader.build_harvester",
             mock_build_harvester,
         )
